requirements.txt
matplotlib
numpy
opencv_python
opencv_python_headless
pandas
Pillow
pynput
